pyproject.toml
[tool.poetry]
name = "streamrss"
version = "0.1.0"
description = "RSS feed streamer"
authors = ["0xPeterSatoshi <0xpetersatoshi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
feedparser = "^6.0.10"
sseclient = "^0.0.27"
click = "^8.1.3"
fastapi = {extras = ["all"], version = "^0.79.0"}
sse-starlette = "^0.10.3"
asyncio = "^3.4.3"
beautifulsoup4 = "^4.11.1"
html5lib = "^1.1"
requests = "^2.28.1"
aiohttp = "^3.8.1"
SQLAlchemy = "^1.4.39"
FastAPI-SQLAlchemy = "^0.2.1"
alembic = "^1.8.1"
psycopg2 = "^2.9.3"
python-dotenv = "^0.20.0"
nats-py = "^2.1.4"
asyncio-nats-streaming = "^0.4.0"
pytz = "^2022.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
streamrss = "streamrss.main:main"
worker = "streamrss.feeds:main"